current TODO: Implement SPH smoothing kernel W(r, h) = (4 / (π * h^8)) * (h^2 - r^2)^3   for 0 <= r <= h Add density claculation Add pressure forces Implement spatial hashing / grid based neighbour search move sim to Fixed update Compute shaders for performance?

+++ CFD/Assets/ParticleManager.cs
@@ -7,7 +7,7 @@ public class ParticleManager : MonoBehaviour
 
 
 
-  public float particleCount;
+  public int particleCount;
 
   void Update()
   {
